backend/requirements.txt
Flask==3.1.0
SQLAlchemy==2.0.36
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.7.1
Flask-CORS==5.0.0
marshmallow==3.23.2
gunicorn==23.0.0
python-dotenv==1.0.1
Werkzeug==3.1.3
openpyxl==3.1.5
